Separated AnimationGroupAPI from Position source file. Implemented AnimationGroupAPI quickTo.
